feat: typescript 进度缓慢
