+++ server/index.js
@@ -38,11 +38,11 @@ router.post('/upload', async (ctx) => {
 	console.log(ctx.request.files);
 	const file = ctx.request.files.file;	// 获取上传文件
 	// console.log(file)
-	const reader = fs.createReadStream(file.path);	// 创建可读流
-	const ext = file.name.split('.').pop();		// 获取上传文件扩展名
+	// const reader = fs.createReadStream(file.path);	// 创建可读流
+	// const ext = file.name.split('.').pop();		// 获取上传文件扩展名
 	// console.log(ext)
-	const upStream = fs.createWriteStream(path.resolve(__dirname, `./upload/${Math.random().toString()}.${ext}`));		// 创建可写流
-	reader.pipe(upStream);	// 可读流通过管道写入可写流
+	// const upStream = fs.createWriteStream(path.resolve(__dirname, `./upload/${Math.random().toString()}.${ext}`));		// 创建可写流
+	// reader.pipe(upStream);	// 可读流通过管道写入可写流
 	return ctx.body = '上传成功';
 })
