--- server/temp/upload_3a5fe00a21f1b46f791f67410692d75c
@@ -1,81 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<svg
-   xmlns:dc="http://purl.org/dc/elements/1.1/"
-   xmlns:cc="http://creativecommons.org/ns#"
-   xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-   xmlns:svg="http://www.w3.org/2000/svg"
-   xmlns="http://www.w3.org/2000/svg"
-   xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd"
-   xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape"
-   width="248.22169"
-   height="67.088432"
-   viewBox="0 0 248.22169 67.088432"
-   version="1.1"
-   id="svg3760"
-   sodipodi:docname="filestack.svg"
-   inkscape:version="0.92.1 r">
-  <metadata
-     id="metadata3764">
-    <rdf:RDF>
-      <cc:Work
-         rdf:about="">
-        <dc:format>image/svg+xml</dc:format>
-        <dc:type
-           rdf:resource="http://purl.org/dc/dcmitype/StillImage" />
-        <dc:title>color</dc:title>
-      </cc:Work>
-    </rdf:RDF>
-  </metadata>
-  <sodipodi:namedview
-     pagecolor="#ffffff"
-     bordercolor="#666666"
-     borderopacity="1"
-     objecttolerance="10"
-     gridtolerance="10"
-     guidetolerance="10"
-     inkscape:pageopacity="0"
-     inkscape:pageshadow="2"
-     inkscape:window-width="3190"
-     inkscape:window-height="1762"
-     id="namedview3762"
-     showgrid="false"
-     inkscape:zoom="4.1781768"
-     inkscape:cx="118.33668"
-     inkscape:cy="33.346753"
-     inkscape:window-x="0"
-     inkscape:window-y="28"
-     inkscape:window-maximized="0"
-     inkscape:current-layer="svg3760"
-     inkscape:lockguides="false" />
-  <!-- Generator: Sketch 42 (36781) - http://www.bohemiancoding.com/sketch -->
-  <title
-     id="title3742">color</title>
-  <desc
-     id="desc3744">Created with Sketch.</desc>
-  <defs
-     id="defs3746" />
-  <g
-     id="Page-1"
-     style="fill:none;fill-rule:evenodd;stroke:none;stroke-width:1"
-     transform="scale(1.3993862)">
-    <g
-       id="color">
-      <path
-         d="M 37.498873,32.959665 V 4.4944997 H 4.4998648 V 43.446831 H 26.999189 V 47.94133 H 41.998738 V 32.959665 Z M 0,0 H 41.998738 V 47.94133 H 0 Z M 29.999099,35.955998 H 41.998738 L 29.999099,47.94133 Z M 13.499594,13.483499 h 14.99955 v 4.4945 h -14.99955 z m 0,7.490833 h 11.99964 v 4.4945 h -11.99964 z m 0,7.490833 h 4.499865 v 4.4945 h -4.499865 z"
-         id="Rectangle-356"
-         inkscape:connector-curvature="0"
-         style="fill:#ef4925" />
-    </g>
-  </g>
-  <text
-     xml:space="preserve"
-     style="font-style:normal;font-variant:normal;font-weight:normal;font-stretch:normal;font-size:33.33333206px;line-height:1.25;font-family:Montserrat;-inkscape-font-specification:'Montserrat, Normal';font-variant-ligatures:normal;font-variant-caps:normal;font-variant-numeric:normal;font-feature-settings:normal;text-align:start;letter-spacing:0px;word-spacing:0px;writing-mode:lr-tb;text-anchor:start;fill:#000000;fill-opacity:1;stroke:none"
-     x="74.019226"
-     y="46.381126"
-     id="text3700"><tspan
-       sodipodi:role="line"
-       id="tspan3702"
-       x="74.019226"
-       y="46.381126"
-       style="font-style:normal;font-variant:normal;font-weight:normal;font-stretch:normal;font-size:33.33333206px;font-family:Montserrat;-inkscape-font-specification:'Montserrat, Normal';font-variant-ligatures:normal;font-variant-caps:normal;font-variant-numeric:normal;font-feature-settings:normal;text-align:start;writing-mode:lr-tb;text-anchor:start">filestack-js</tspan></text>
-</svg>
