+++ src/lib/api/upload/index.ts
@@ -1,3 +1,1 @@
-
 export * from './upload';
-export * from './types';
