+++ src/lib/api_bak/request.ts
@@ -1,0 +1,26 @@
+import * as agent from 'superagent';
+
+/**
+ * @private
+ */
+export interface CustomReq extends agent.SuperAgentStatic {
+  [method: string]: any;
+}
+
+/**
+ *
+ * @private
+ * @param method
+ * @param url
+ */
+const requestWithSource = (method: string, url: string): CustomReq => {
+  const newReq: CustomReq = agent;
+  return newReq[method](url).set('Filestack-Source', 'JS-@{VERSION}');
+};
+
+/**
+ * @private
+ */
+const request: agent.SuperAgentStatic = agent;
+
+export { request, requestWithSource };
