+++ src/lib/api_bak/upload/index.ts
@@ -1,1 +1,3 @@
+
 export * from './upload';
+export * from './types';
