+++ src/lib/core.ts
@@ -1,10 +1,12 @@
-import { upload, simpleUpload } from './api/upload';
-import TaskManager from './TaskManager';
+import TaskManager, { Task, Tasks } from './TaskManager';
 import { EMimeType } from '../types/upload';
 import Logger from './logger';
 import { ChunkTask } from './task/ChunkTask';
 import DirectTask from './task/DirectTask';
-
+import { UploadConfig, UploadOptions } from './api/upload/types';
+import BaseTask from './task/BaseTask';
+import Config from '../config/index';
+import { AxiosResponse, AxiosError } from 'axios';
 export interface FilePortalOptions {
   debug: boolean;
   chunkSize: number;
@@ -15,23 +17,46 @@ export interface FilePortalOptions {
 }
 
 export interface AddOptions {
-  immediate: boolean;  // 加入后立即上传
+  // immediate: boolean;  // 加入后立即上传
+  token: () => Promise<string> | string;
+  extra?: Object;
+  config: UploadConfig;
 }
 
-export class FilePortal extends TaskManager {
+export interface EventResponse {
+  complete?: ((res?: Tasks) => any);
+  error?: ((err?: any) => any);
+  uploaded?: (tasks: Tasks, task?: Task) => any;
+}
+
+export const enum Events {
+  COMPLETE = 'complete',  // 所有task完成
+  UPLOADED = 'uploaded',  // 某个task完成
+  ERROR =  'error',       // task出错
+}
+
+const noop = () => {};
+
+export default class FilePortal {
   debug;
   options: FilePortalOptions;
   taskManager: TaskManager;
+  events: EventResponse;
   constructor(options?: FilePortalOptions) {
-    super();
+    // super();
     this.debug = new Logger('filePortal/core');
     this.options = options;
-    this.taskManager = new TaskManager();
+    this.taskManager = new TaskManager(this);
+    this.events = {
+      complete: noop,
+      error: noop,
+      uploaded: noop,
+    };
   }
 
   static BLOCK_SIZE = 4 * 1024 * 1024;
 
-  private generateTask = (file: File) => {
+  private generateTask = (file: File): BaseTask => {
     if (file.size > FilePortal.BLOCK_SIZE) {
       return new ChunkTask(file, FilePortal.BLOCK_SIZE, 2 * 1024 * 1024);
     } else {
@@ -39,18 +64,63 @@ export class FilePortal extends TaskManager {
     }
   }
 
-  add(file: any, options: AddOptions) {
+  add(file: any, options: AddOptions): Tasks {
     this.debug.log('upload start');
     /* istanbul ignore next */
-    return upload(null, null, null);
+    let task: BaseTask = this.generateTask(file);
+    return this.taskManager.addTask(task, options);
   }
-  simpleUpload() {
-    return simpleUpload('keke', {
-      config: {
-        host: 'localhost:8089',
-      },
-    } as any);
+
+  start(tid) {
+    return this.taskManager.start(tid);
   }
+
+  pause(tid) {
+    this.taskManager.pause(tid);
+    return this;
+  }
+
+  resume(tid) {
+    this.taskManager.resume(tid);
+    return this;
+  }
+
+  cancel(tid) {
+    this.taskManager.cancel(tid);
+    return this;
+  }
+
+  complete(cb: (tasks?: Tasks, task?: Task, source?: string) => any) {
+    this.events.complete = cb;
+  }
+
+  uploaded(cb: (tasks?: Tasks, task?: Task, source?: string) => any) {
+    this.events.uploaded = cb;
+  }
+
+  error(cb: (err?: any) => any) {
+    this.events.error = cb;
+  }
+
+  on(evt: string, cb): void {
+    if (evt === Events.COMPLETE) {
+      this.complete(cb);
+    }
+    if (evt === Events.ERROR) {
+      this.error(cb);
+    }
+    if (evt === Events.UPLOADED) {
+      this.uploaded(cb);
+    }
+    return;
+  }
+  // simpleUpload() {
+  //   return simpleUpload('keke', {
+  //     config: {
+  //       host: 'localhost:8089',
+  //     },
+  //   } as any);
+  // }
   setOptions(options: FilePortalOptions) {
     this.options = {
       ...this.options,
