+++ tsconfig.json
@@ -26,7 +26,7 @@
       ]
     },
     "include": [
-      "src/**/*.ts"
+      "src/**/*.ts", "src/lib/__test__/core.test.js"
     ],
     "exclude": [
       "node_modules/**"
